Separate and export BuildReceiverIntegrations (#3553)
* Move and export BuildReceiverIntegrations


---------
